requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-Mail==0.9.1
librosa==0.10.0
python-dotenv==1.0.0
Werkzeug==2.2.3
supabase==1.0.3
